change percentage to progress, merge only progress, remove camera state code that didn't work
